requirements.txt
google-generativeai==0.1.0
prompt_toolkit==3.0.36
python-dotenv==0.19.2
requests==2.28.1
